Merge cmake-build with the master branch

Former-commit-id: 9012f0558111a3a9261f863b6332e8c44d02003f [formerly ac812a50a6da8c6424481b62c6556ea7f1e91a57 [formerly d417326f4f6b502bc77934dbfc50995267ff458e]]
Former-commit-id: 664704f434fb3bb7de1a796210703b240bc8dc59
Former-commit-id: d8a652e534a08db336170d89f0a96a5e2cc8de41

+++ src/game/ui/Navball.cpp
@@ -1,3 +1,5 @@
+#include <cmath>
+
 #include "Navball.h"
 
 #include "../../util/DebugDrawer.h"
@@ -31,7 +33,7 @@ void Navball::draw_to_texture(const Vessel& vessel, const ReferenceFrame& frame)
 	// we know the sine as we know the opposite to the angle (1.0f) and the
 	// hypothenuse (view_distance)
 	float angle_sin = 1.0f / view_distance;
-	float fov = std::asinf(angle_sin) * 2.0f;
+	float fov = std::asin(angle_sin) * 2.0f;
 
 	glm::mat4 view, proj, model;
 	// 60 degrees as we see the sphere from twice its radius
